[TML/BC/OGS] Move CellInformation to helpers
